requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.15
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-embed-video==1.3.3
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.10
Pillow
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
